Split off golden file exceptions from main logic
